pyproject.toml
[project]
name = "autochecklist"
version = "0.1.0"
description = "A library of checklist generation and scoring methods for LLM evaluation"
readme = "README.pypi.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pydantic>=2.0",
    "httpx>=0.25",
    "python-dotenv>=1.0",
    "tenacity>=8.0",
    "tqdm>=4.0",
]

[project.urls]
Homepage = "https://github.com/ChicagoHAI/AutoChecklist"
Repository = "https://github.com/ChicagoHAI/AutoChecklist"
Documentation = "https://autochecklist.github.io"
"Bug Tracker" = "https://github.com/ChicagoHAI/AutoChecklist/issues"

[project.optional-dependencies]
ml = [
    "numpy>=2.2",
    "networkx>=3.4",
    "transformers>=4.40",
    "datasets>=4.5",
]
vllm = ["vllm>=0.8.5"]
notebooks = ["ipykernel>=7.1", "nest-asyncio>=1.6"]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-autoapi>=0.4.1",
    "mkdocs-material>=9.7.1",
    "mkdocstrings[python]>=1.0.3",
]
all = ["autochecklist[ml,vllm,notebooks]"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
]

[project.scripts]
autochecklist = "autochecklist.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["autochecklist"]

[dependency-groups]
dev = [
    "autochecklist[ml]",
    "ipykernel>=7.1.0",
    "mkdocs>=1.6.1",
    "mkdocs-autoapi>=0.4.1",
    "mkdocs-material>=9.7.1",
    "mkdocstrings[python]>=1.0.3",
    "nbconvert>=7.17.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
markers = [
    "integration: makes real API calls (skip with -m 'not integration')",
    "vllm_offline: requires vLLM and GPU for offline inference tests",
    "openai_api: requires OPENAI_API_KEY for direct OpenAI API tests",
]
testpaths = ["tests"]
